Moving API controllers test to separate file (task #5386)
